requirements.txt
pandas
fastapi
strawberry-graphql
pydantic
uvicorn